Added menu links, CarController, cars.html, ConvertService, repair_invoice_form.html, repair_invoices.html, RepairInvoiceController, RepairInvoice class, dto, repository, Modified schema.sql, CarRepository query, CarService, CarStatusService, OrderController, OrderRepository, OrderService,OrderStatusService, SecurityConfiguration, SqlConstant, UserDetailsService, users.html

+++ repository-module/src/main/java/com/gmail/silina/katsiaryna/repository/OrderStatusRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderStatusRepository extends JpaRepository<OrderStatus, Integer> {
+public interface OrderStatusRepository extends JpaRepository<OrderStatus, Long> {
     @Query("SELECT os FROM OrderStatus os WHERE os.orderStatusEnum = :status")
     OrderStatus findByStatus(@Param("status") OrderStatusEnum status);
 }
